AP-7033: Fixed a bug occuring with abbrevations
Fixed a bug where the dot of an abbrevations which is at the end of a
sentence would have been copied to be a single term as well.
Example: This is Knime Corp. -> Terms: This, is, Knime, Corp., .
The copied dot breaked the DocumentBuilder().
Additionally the .ancoraFactory() method has been replaced, since the
options coming with the ancoraFactory are wrong and would have been
overwritten anyway.

ERNEUT GEÖFFNET: AP-7033 (Text Mining: Default normalization of Spanish
tokenizer need to be switched off)
https://knime-com.atlassian.net/browse/AP-7033